+++ NEWS.md
@@ -1,3 +1,11 @@
+# simStateSpace 1.2.16
+
+## Patch
+
+* Added the `SimMVN()` function.
+* Consolidated the `SimAlphaN()`, `SimIotaN()`, `SimNuN()`, and  `SimMuN()` functions.
+* Added the `LinSDEInterceptEta()` and `LinSDEInterceptY()` functions.
+
 # simStateSpace 1.2.15
 
 ## Patch